Remove the #exit on push failures
Unless I have missed something in the code, it seems that this exit
command would kill the entire daemon when there is a push error in
only one of the runners. That seems overkill.

However, I have added a TODO marker to go back and record status
information about the error. This could then be displayed through
either the web interface or the CLI status.

This should fix #47