Rescue any StandardErrors in #push_changes
The commands involved in the push and end up throwing errors that are
not critical, or will be resolved in the next check.
The method has been updated to rescue all StandardErrors and trigger a
notification. Similar to the previous commit, errors in a single share
should probably not kill the entire daemon.

Also similarly to the last commit these types of errors could probably
be recorded as status information on the share which was affected.

This should fix #53