Re-write empty directory tagging to use Find
Find is a module in the stdlib which will talk through a directory tree.
The code will now flag empty hidden directories in a repository while
correctly ignoring the .git directories. (The tests has also been
updated to validate this.)

I am hoping that this will also improve the performance during a commit
but that has not yet been proved.

Additionally, I also added a test helper method for creating directories
in the local testing repository.