Stop the addition of extra conflict markers
Due to a mis-understanding of how the Rugged index_entry list works,
and not having a test which covered this particular scenario the
conflict marking was marking all of the files during a conflicting merge
instead of just the conflicted ones.

The code has been modified to specifically look for index_entries which
have 2 or more conflicting commits. A test case covering this scenario
has also been added to the test suite.

I think this will resolve issue #82