Extract push and merge notifications into Notifier
In the process of making some changes to the Runner it was being made more
complicated by having to account for the different possible notifications.

And this pain made be re-think the design a little bit. The notification
sections do not have anything to do with the rest of the runner, so
extracting was straight forward.