Merge push_changes into sync_changes
When doing a sync on the repository we will now always follow the order:
* commit
* fetch
* merge
* push
This will help to ensure that changes are quickly committed and reduce
conflicts when pushing to the repository.

This also simplifies the code because we no longer need #push_changes,
as #sync_changes is always called.