Extract update and delete from the Server
I have extracted the core of these calls, which is manipulating the
configuration out into the Configuration class. The methods are:
* remove_by_id
* update_all

I have also added delegates to the Manager class, so that the Server
does not need to know that the config object is used. This eliminates a
LawOfDemeter violation.