Extract text mime type check into Path
Since we are only ever checking if the file is text, I am extracting
that out into the Repository::Path instead of letting the Server know
the details of what makes a text file.

This should also ensure that the check for a text file is consistent
where ever it is done.