Fix a Javascript error on the revisions page
I probably introduced this bug during the Rugged re-writes.

The Javascript assumes dates are formatted as ISO8601, which it got in
the regular file listing (because the data passed through #to_json which
formats dates as ISO8601) but not in the revision list. (where the dates
were getting displayed with an implicit #to_s)

Adding a #iso8601 to the date in the revision view seems to resolve this
error.