Convert web UI to use Sinatra
In the process I split it into 2 apps: one for the settings, and one
for browsing the repositories.

I adjusted the HAML templates a little to account for Sinatra.
And moved the default layout to 'layout.haml' which is the default for
Sinatra.

The renee dependency has been remove, since it is not longer needed.