Add code fencing and syntac highlighting
I think that this fixes #41. Code fencing is not enabled in the
Redcarpet Markdown parser and that code can be syntax highlighted using
the CodeRay highlight we is already included.

I found and example of how the pass CodeRay into Redcarpert on
StackOverflow:
https://stackoverflow.com/questions/17830065/code-syntax-highlighting-with-coderay

And then did some refinement on the original idea:
* allow fenced code to work with or without a specified language
* correctly pass the options through Tilt
* extract the rendering of content into a Sinatra helper

While making these modifications also discovered 2 bugs in BrowserApp:
* inline README would failed because I mis-spelled #gsub
* embedded file display was looking at the incorrect raw address

These are both resolved in the updated code.