Extract Share and Initializer from Configuration
The original motivation for this refactoring is to reduce the number of
objects which are getting passed around making the Celluloid conversion
easier. ActiveRecord objects (e.g., Share and Config) will be used from
their class methods, more like Rails does. This will eliminate some
object passing as well as deprecating some delegate code in the
Configuration and Manager classes.

This results in 3 classes instead of 1 and I hope will make each one
easier to understand, and test.

I have kept Configuration::Config to avoid changing the table names.
This could be simplified in the future.