Convert BrowserApp to use ID instead of index
This allow the Share.find to be used instead selecting by index. This
might be more performant, but my main reason for the change is that
using the database ID is more expected than using the index.

Also removed idx from getting passed into the views because it is no
longer being used.

(N.B.: There is no change to the SettingsApp because it works on all of
the shares at once, and does not care about the IDs.)