Convert to Celluloid
I have replaced event_machine/thin with celluloid/listen/reel.
(Eventually this will allow the guard dependency removed, but the
notifications have not yet been converted to the latest version.)

The only functional change is that the new Listen gem does not seem to
get notified of the creation of empty directories. We can probably live
with this. Otherwise, I have attempted to keep the execution the same.
(Celluloid should allow for some future improvements.)o

Supporting changes:
* removed class which are no longer used
* added additional query methods to Gitdocs::Share for finding
  appropriate sets of shares