Upgrade rugged to v0.23.0
Update branch and remote access to use the new rugged API.

In the integration tests I needed to update how the rugged gem is
required to ensure that it will respect the value of HOME which is set
for the test. (rugged/libgit2 appear to read HOME on load and never
update the value being used)

In changing how rugged is required, I also eliminated loading most of
the gitdocs code in the integration test runner, because it executes
the test through the CLI and web UI. The remaining ones are marked with
a FIXME flag, and can be removed once the CLI supports some additional
more configuration settings. (e.g., disabling notifications and the
polling interval)