Replace coveralls with CodeClimate coverage
The coveralls gem has a dependency on the tins gem which requires
Ruby v2+ and prevents TravisCI from running the tests in Ruby1.9.
But since we are already using CodeClimate we can switch to its code
coverage tool and eliminate the coveralls dependency.

I have also updated the badges in the README. (And updated all the
badges to use SVG instead of PNG.)