Fix clear command missing pid and restart
I noticed this when the 'clear' integration test started failing
because it was trying to send a pid argument, which was being rejected.
(I am not sure when this argument stopped being ignore by the clear
command. Maybe an upgrade to the Thor gem?)

While adding the 'pid' argument I also realized that the 'clear'
command was not resenting the running daemon. Now it does, and the
integration tests work too.