Comment on limiting activerecord upgrades
If activerecord is upgrade to v5.x then we lose support for Ruby v2.0.
Which would be OK except that the default version of run in OSX is
still v2.0. So even though it is EOL, it seems appropriate to retain
support for it for the moment.

I have added a comment into the gemspec to indicate why activerecord
should not be upgraded to v5.x until this becomes different.